Refactor HCloudClient using a factory approach
In order to facilitate testing where we need to have fake HCloudClients,
we refactor the cluster scope as well as the ClusterReconcile object to
include a HCloudClientFactory. This factory will be initialized in
main.go (or in future test suites). Instead of reading directly from the
secret in scope, we get the HCloud token from now on in the main
reconcile function, which makes the structure much clearer.