:bug: Fix wrongly initialized slice of services
As the load balancer services in the API include one service we call
the kubeAPI service, we always have one service too much in the list. We
don't count this kubeAPI service in our list of ExtraServices, so that
we have to always consider slices of length len(lb.Services)-1.