:sparkles: Size in GB in hardware details
As the size of a storage device is very difficult to read for a human
eye if it is written in bytes, we add the size in GB to the details of a
storage device.

Additionally, there are a few bug fixes in the tests, mostly related to
the hostname prefix.