:sparkles: Bare metal control planes
In order to support bare metal control planes, they have to be added and
removed as targets to/from the control plane load balancer. For this, we
need to support adding and removing IP targets to/from the load
balancer.

These different target types (before we only used type server), requires
an API change. Load balancer targets cannot be a slice of (server) IDs,
but need to reflect the different types (server and IP) that are now
possible.

One main issue was to align providerIDs, so that this provider
integration works well together with the cloud controller manager.

In order to test this new bare metal only set-up, there is a new button
in Tilt to start such a workload cluster.