:bug: Add server to lb with only private IP
In one of the last PRs we added support for Hetzner's new feature of
disabling IPv4 and IPv6 addresses. This required to prevent servers from
being added to load balancer in case they only have IPv6 addresses - as
no IPv6 addresses can be added to a Hetzner load balancer. This only
applies to public network though. This is why we also need to add
servers in case that private networks are there - independent whether
there is a public IPv4 or not.