:bug: Add kubeadm reset to deprovision
At the moment, the cloud controller manager continues to run on servers
that have been deprovisioned. This causes problems with rate limits
almost immediately, as it continues to retry requests that return
errors. For this, we add an ssh command to reset kubeadm.

Additionally, there is a small issue with ssh key names that are
uploaded to Hetzner. If they contain a line break at the end, the code
fails to recognize that it actually is the same secret. We trim line
breaks now.