:bug: Fix enabling IP address logic for HCloud
Currently, the user can only enable/disable IPv4 and IPv6 if there is no
private network. However, since the enabling/disabling is independent of
the public/private net, we remove this if clause.

To make sure that users don't disable the IPv4 if there is no private
net, we always enable it in that case. This is necessary for Hetzner
load balancers, as they can only work with IPv4 of HCloud servers.

Fix #312