:bug: Prevent secret changed error on cluster move
When a bare metal cluster is moved, the resource versions of the secrets
change. However, this leads to an error message in all provisioned bare
metal servers, as they expect that the ssh key that they have been
provisioned with is not there anymore.

Instead of looking at the resource version, we compare the data of the
secret with hash values. If nothing changed there, we expect the secret
(or rather the ssh keys stored in it) to still be the same.

Fixes #461