:bug: Fix ssh call to wrong port in actionEnsureProvisioned
In actionEnsureProvisioned we check whether cloud init actually ran
successfully: Sometimes it happens that it shows success, but has a
SIGTERM in the logs.

However, if we change ssh ports in cloud init, then we have to call the
new port after cloud init finished. If this works, we can already
conclude that cloud init did something right. This is enough reason for
us to assume that everything went well.

If this is a wrong assumption, we can include in the future an additional if statement
that makes ssh statements to both ports to check the current cloud init
status.