:sparkles: Add remediation controller for hcloud
A remediation controller that reboots hcloud servers for a specified
amount of times before it deletes and recreates the machine. This
pattern is more efficient in the majority of cases where HCloud servers
get healthy again after a rebootThis pattern is more efficient in the
majority of cases where HCloud servers get healthy again after a reboot.