:sparkles: Add conditions to HetznerBareMetalHost
Adding conditions in case that something is wrong with secrets or
credentials which the user has to specify. There are secrets for Hetzner
robot and for the SSH keys used to provision the bare metal servers.

Removing the error message of spec.status from printcolumns, as it is
used only internally for the controller to manage the provisioning
process.