:seedling: Increase e2e timeout for control planes
Our timeout for control planes is currently 15 minutes. This is not
enough for bare metal, as it is used separetely for a) the first control
plane (which needs max 15m) and b) all the rest of the control planes
(which need more time from the point where the first control plane is
ready).