:seedling: Improve error handling for locked servers
When a HCloud server is locked, this is because the API is doing
something for a short time where no other API request for that server
can go through - namely right after a server is created. We ignore these
errors in the future and just trigger another reconcilement.

Additionally, some unused log statements are removed.