:sparkles: Get hcloud server by ID and labels
To reduce the number of expensive API calls, the findServer function
first tries to call GetServer with the serverID that is stored in the
providerID of the HCloudMachine object.

If the call fails or the ProviderID is not set, the server will be
queried via labels - as was the case until now.