:bug: Keep finalizer on secrets with multiple clusters
In the case where there are multiple clusters referencing on the same
secret, every cluster that gets deleted removes the finalizer from the
secret. This does not make sense if there are other clusters also
referencing the secret. This is why we keep the finalizer on secrets as
long as there are other owner references on the secret when a cluster
gets deleted.

Additionally, we remove the owner reference of the HetznerClusters from
the secret when deleting a cluster. This means that the secrets don't
get deleted, even if they don't have any other owners.