fix: reduce reconciles by sorting networkStatus.attachedServers
The list of servers returned by the API is not consistently sorted.
As the IDs from this list are written into
HetznerCluster.Status.NetworkStatus.AttachedServers in this random
order, this causes uncesseary patches to the HetznerCluster object.

By sorting the IDs before writing, only actual changes causes the object
to be patched.

As other controllers have a watch on HetznerCluster, this reduces the
general amount of reconciles happening in CAPH.