:bug: Add kubeapi service to existing load balancer
The feature of taking over existing load balancers did not pay attention
to the kubeapi service which is not added to the load balancer. This is
due to the fact that this was done in the creation process only.

This commit adds it to the reconcilement of the extraServices so that it
gets added to an existing load balancer as well.