:sparkles: Try ssh reboot in rescue
If a bare metal server is provisioned for the second time after a first
successful provisioning, and as long as the ssh key did not change, we
can still access it with ssh and can perform an ssh reboot instead of an
API reboot (sw, hw). In this way, we can save some API calls.

This commit introduces this change.

It further fixes two small things where misleading or unnecessary logs
were printed.