:seedling: Improve startup of target cluster manager
Improving the startup of the target (workload) cluster manager to avoid
unnecessary and temporary errors that come while the first control plane
is still provisioning.

This is done by an additional check whether the control plane is
available already.

Furthermore, this commit introduces a new condition so that the Hetzner
secret that is stored in the target cluster reconciles with an own
condition.