:bug: Fix provision condition time lag
There has been a time lag in the condition that shows the provisioning
status of the HetznerBareMetalHost. It was only updated after a new
reconcilement has started, not during the reconcilement where the state
actually changed.

This commit places the update of the condition to the start of the
respective functions to ensure it is up to date.