:seedling: Add event if host is provisioned
Adding an event to show that the provisioning process has been
completed. This is important to find out based on events where an issue
happens. Even if it is successfully provisioned, it might not be added
to the cluster, as the CCM does some work after CAPH. To know more about
the state, this event will be helpful.