:bug: Fix wrong names of storage devices (#1130)
Names of storage devices might change after reboot. We stored the names
in the host object and didn't update them anymore. However, these names
might change after a reboot and might even switch between devices.

This led to the issue that the wwn that the users give as reference to
the storage device that is used in provisioning did not correspond
anymore to the name of the device that we stored in the host object.

This led to the operator using the wrong storage device in provisioning.

Now we get the latest information of wwn and storage device names before
installimage and thus always take the correct storage device for
provisioning.

Additionally, we don't store the names of storage devices in the host
object anymore, because they might be outdated and confuse users.