fix links failures and enable link checking in docs directory
when we mark a PR ready_for_review and if the changes are in docs
directory then we are not checking for link failures and this was the
reason why there was a commit on main which introduced link failures.

This commit fixes the same and introduces link checking even if we do
some changes in the docs directory.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>