:bug: MachineDeployment with RAID, reject host if invalid WWN config.
If a HetznerBareMetalHost contained invalid WWN configuration (single WWN and RAID WWN),
then the host was chosen for provisioning. Before provisioning we did not check
again, so that it could happen that a host with one WWN got provisioned for a
MachineDeployment with RAID. This resulted in a not easy to understand
"installimage failed" error.

Co-authored-by: Yash Raj Singh <yash.raj@syself.com>