:sparkles: Add ssh key name via hetzner secret
Adding the ssh key name via hetzner secret facilitates use cases where a
lot of clusters are created with the same ssh key. Instead of having to
specify it every time in the HetznerClusterTemplate object, it can be
specified in the secret ones.

On top, this commit introduces default values for the HetznerSecret to
allow use cases where all is hard-coded and doesn't have to be specified
e.g. in a ClusterClass or in the templates of the cluster object
anymore.