:seedling: handle case where rescue system can't be reached. (#1220)
Provide better err msg if rescue system can't be reached. Handle this case better, because currently it reaches and endless loop of hardware reboots if a bare metal server's rescue system is not reachable.

Additionally, the controller emits an Event for every reboot now. This makes it more clear what is happening.