:sparkles: Add ssh keys to status and fix ssh key in secret
In order to visualize which ssh keys have been used to provision a
machine, we write the keys in the status of the HCloudMachine.

On top, we fix a bug where the "patch" call for the finalizer removed
the ssh key from the spec of the HetznerCluster again, because it cannot
patch the HetznerCluster.spec if a ClusterClass is used.

We add the ssh key now after the patch, so that it will still get
patched at the end of the function, but can be used until then. In case
of a ClusterClass we just don't have the key reflected in the spec of
the HetznerCluster.

This is why we also add the list of keys to the HCloudMachine status, so
that it is better visible.