:seedling: Stop calling API when HCloudMachineTemplate status is set (#1353)
:seedling: Stop calling API when hcmt status is set

We call the HCloud API in every reconcile loop of the
hcloudmachine_template controller. This is not necessary since the
server type does not change for one template object. We can set it once
when the status is empty and then don't have to set it while the status
is not nil anymore.