:bug: Increase timeout on connection refused errors
We have the timeout of one minute that we allow connection refused
errors to come after reboot of a server.

In this commit we added the (correct) permanent error if we conclude the
reboot to be "failed". However, apparently, for a long time already, we
marked reboots as failed that did not actually fail.

This commit introduces a larger timeout, so that we don't mark the
reboots as failed, and in turn don't mark the HetznerBareMetalHost as
failed.

We also increase the timeouts to a minimum of 10 min, so that we don't
run into failed states so quickly