:bug: Remove false duplicate validation in webhooks  (#1395)
:bug: Remove false duplicate validation in webhooks

We introduced a bug previously by adding a duplicate validation in the
webhooks of HCloudMachineTemplate and HetznerBareMetalMachineTemplate in
the update function. There we used the function
ShouldSkipImmutabilityCheck that allows to use the webhook with Cluster
API's ClusterClass. However, we added another function call that again
checks that everything is immutable and didn't use the above utility
function from Cluster API. That's why the webhooks were not compatible
with ClusterClass anymore. All we have to do is to remove the duplicate
call that was not needed anyway.