:seedling: Remove permanent error after bm reboot
In the provisioning of bare metal servers, we added permanent errors
that appear after a server times out while rebooting.

This turned out to be quite unreliable, because the server reboots are
very unreliable. Sometimes they take forever to reboot, in other
situations they are fast and the server provisions successfully.

Therefore, we should rather try again and not stop completely.