:seedling: Install Cloud-Init-Data via post-install.sh (#1407)
* :seedling: Copy Cloud-Init-Data via post-install.sh (bare-metal)

This PR copies the cloud-init files to the future root-filesystem from the rescue system. This eliminates the State "Provisioning" and this means machines need only 2 (no more 3) reboots: The first reboot enters the rescue system, the second boot starts cloud-init and the operating system.

This makes provisioning bare-metal machines faster.

Additionally: If install-image fails, the process could have reached an infinite-loop of re-trying install-image up to now. The PR switches the machine back to "registering" if install-image failed.